+++ js/get-data.js
@@ -1,3 +1,9 @@
+
+
+import {
+  getRandomInt, getRandomIntRange, isCount
+} from './util.js';
+
 const data = {
   comments: [
     'Эй, цирюльникъ, ёжик выстриги, да щетину ряхи сбрей, феном вошь за печь гони!',
@@ -92,6 +98,40 @@ const data = {
   ]
 };
 
+const createId = getRandomIntRange(1, 25);
+const createCountLike = getRandomIntRange(15, 200);
+
+const createCountDescription = getRandomIntRange(0, 24);
+const createTextDescription = () => data.description[createCountDescription()];
+
+const createNamePhothos = getRandomIntRange(1, 25);
+const createUrlPhothos = () => `img/${createNamePhothos()}.jpg`;
+
+const createUrlAvatar = () => `img/${getRandomInt(1, 6)}.svg`;
+const createTextComments = () => data.comments[getRandomInt(1, 6)];
+const createCommentsId = isCount();
+
+const createComments = () => ({
+  id: createCommentsId(),
+  avatar: createUrlAvatar(),
+  message: createTextComments(),
+  name: data.name[getRandomInt(0, 29)]
+});
+
+const createObjectTextComments = () =>
+  Array.from({
+    length: getRandomInt(0, 29)
+  }, createComments);
+
+
+const getData = () => ({
+  id: createId(),
+  likes: createCountLike(),
+  url: createUrlPhothos(),
+  description: createTextDescription(),
+  comments: createObjectTextComments()
+});
+
 export {
-  data
+  getData
 };
