--- js/getRandomInt.js
@@ -1,3 +1,0 @@
-const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-export { getRandomInt };
