--- js/isCount.js
@@ -1,11 +1,0 @@
-const isCount = () => {
-  let count = 0;
-
-  return function () {
-    return count++;
-  };
-};
-
-export {
-  isCount
-};
