+++ js/main.js
@@ -1,9 +1,9 @@
-import { createObjectPhotos } from './createObject.js';
+import { getData } from './get-data.js';
 
 const COUNT_COMMENTS = 25;
 
-const arrayPhotos = Array.from({ length: COUNT_COMMENTS }, createObjectPhotos);
+const arrayPhotos = Array.from({ length: COUNT_COMMENTS }, getData);
 
-console.table(arrayPhotos);
+// console.table(arrayPhotos);
 
 void (arrayPhotos);
