+++ js/util.js
@@ -1,6 +1,12 @@
-import {
-  getRandomInt
-} from './getRandomInt.js';
+const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+const isCount = () => {
+  let count = 0;
+
+  return function () {
+    return count++;
+  };
+};
 
 const getRandomIntRange = (min, max) => {
   const arrNumberBefore = [];
@@ -22,4 +28,8 @@ const getRandomIntRange = (min, max) => {
   };
 };
 
-export { getRandomIntRange };
+export {
+  getRandomIntRange,
+  isCount,
+  getRandomInt
+};
